[agent v0.3.6] 🐛 fix: kill process group

+++ cmd/agent/main.go
@@ -234,18 +234,32 @@ func doTask(task *pb.Task) {
 		startedAt := time.Now()
 		var cmd *exec.Cmd
 		var endCh = make(chan struct{})
+		var pg utils.ProcessExitGroup
 		timeout := time.NewTimer(time.Hour * 2)
 		if utils.IsWindows() {
+			var err error
+			pg, err = utils.NewProcessExitGroup()
+			if err != nil {
+				// Windows 进程组创建失败，直接退出
+				result.Data = err.Error()
+				client.ReportTask(ctx, &result)
+				return
+			}
 			cmd = exec.Command("cmd", "/c", task.GetData())
+			pg.AddProcess(cmd.Process)
 		} else {
 			cmd = exec.Command("sh", "-c", task.GetData())
+			cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
 		}
-		cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
 		go func() {
 			select {
 			case <-timeout.C:
 				result.Data = "任务执行超时\n"
-				cmd.Process.Kill()
+				if utils.IsWindows() {
+					pg.Dispose()
+				} else {
+					cmd.Process.Kill()
+				}
 				close(endCh)
 			case <-endCh:
 			}
