+++ cmd/playground/main.go
@@ -80,13 +80,19 @@ func cmdExec() {
 		panic(err)
 	}
 	var cmd *exec.Cmd
+	var pg utils.ProcessExitGroup
 	if utils.IsWindows() {
+		pg, err = utils.NewProcessExitGroup()
+		if err != nil {
+			panic(err)
+		}
 		cmd = exec.Command("cmd", "/c", execFrom+"/cmd/playground/example.sh hello asd")
+		pg.AddProcess(cmd.Process)
 	} else {
 		cmd = exec.Command("sh", "-c", execFrom+`/cmd/playground/example.sh hello && \
 echo world!`)
+		cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
 	}
-	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
 	var endCh = make(chan struct{})
 	go func() {
 		output, err := cmd.Output()
@@ -97,8 +103,14 @@ echo world!`)
 	go func() {
 		time.Sleep(time.Second * 2)
 		fmt.Println("killed")
-		if err := syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL); err != nil {
-			panic(err)
+		if utils.IsWindows() {
+			if err := pg.Dispose(); err != nil {
+				panic(err)
+			}
+		} else {
+			if err := syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL); err != nil {
+				panic(err)
+			}
 		}
 	}()
 	select {
